requirements.txt
yfinance
pandas
numpy
matplotlib
pandas-ta
gspread
oauth2client
scikit-learn
python-telegram-bot